@@ -1,4 +1,5 @@
 """Settings for atomate2."""
+from __future__ import annotations
 
 import warnings
 from pathlib import Path
@@ -74,12 +75,12 @@ class Atomate2Settings(BaseSettings):
         None, description="Store data from these files in database if present"
     )
     VASP_STORE_ADDITIONAL_JSON: bool = Field(
-        True,
+        default=True,
         description="Ingest any additional JSON data present into database when "
         "parsing VASP directories useful for storing duplicate of FW.json",
     )
     VASP_RUN_BADER: bool = Field(
-        False,
+        default=False,
         description="Whether to run the Bader program when parsing VASP calculations."
         "Requires the bader executable to be on the path.",
     )
@@ -91,7 +92,7 @@ class Atomate2Settings(BaseSettings):
         "to the simulation will be compressed. If False no file is compressed.",
     )
     VASP_INHERIT_INCAR: bool = Field(
-        True,
+        default=True,
         description="Whether to inherit INCAR settings from previous calculation. "
         "This might be useful to port Custodian fixes to child jobs but can also be "
         "dangerous e.g. when switching from GGA to meta-GGA or relax to static jobs."
@@ -118,7 +119,7 @@ class Atomate2Settings(BaseSettings):
         "cp2k.psmp", description="Command to run the MPI version of cp2k"
     )
     CP2K_RUN_BADER: bool = Field(
-        False,
+        default=False,
         description="Whether to run the Bader program when parsing CP2K calculations."
         "Requires the bader executable to be on the path.",
     )
@@ -149,13 +150,13 @@ class Atomate2Settings(BaseSettings):
         None, description="Store data from these files in database if present"
     )
     CP2K_STORE_ADDITIONAL_JSON: bool = Field(
-        True,
+        default=True,
         description="Ingest any additional JSON data present into database when "
         "parsing CP2K directories useful for storing duplicate of FW.json",
     )
 
     CP2K_ZIP_FILES: Union[bool, Literal["atomate"]] = Field(
-        True,
+        default=True,
         description="Determine if the files in folder are being compressed. If True "
         "all the files are compressed. If 'atomate' only a selection of files related "
         "to the simulation will be compressed. If False no file is compressed.",
